requirements.txt
django>=4.1.0,<4.1.9
djangorestframework==3.14.0
django-cors-headers>=4.0.0,<=5.0.0
six==1.16.0
Pillow==9.5.0
djoser==2.2.0